deck class
