+++ src/test/java/com/github/zipcodewilmington/casino/cards/CardsTest.java
@@ -3,28 +3,37 @@ package com.github.zipcodewilmington.casino.cards;
 import org.junit.Assert;
 import org.junit.jupiter.api.Test;
 import java.util.ArrayList;
+import java.util.Collections;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 class CardsTest {
 
-    @Test
-    public void createDeckTest(){
-        Cards card = new Cards(Rank.ACE, Suit.CLUBS, 14);
-        int expected = 52;
-        ArrayList<Cards> actualCards = card.createDeck();
-        int actual = actualCards.size();
-
-        Assert.assertEquals(expected, actual);
-    }
-
-
-    @Test
-    void testToString() {
-        Cards card = new Cards(Rank.ACE, Suit.CLUBS, 14);
-        int expected = 52;
-        ArrayList<Cards> actualCards = card.createDeck();
-        System.out.println(actualCards);
-    }
+//    @Test
+//    public void createDeckTest(){
+//        Cards card = new Cards(Rank.ACE, Suit.CLUBS, 14);
+//        int expected = 52;
+//        ArrayList<Cards> actualCards = card.createDeck();
+//        int actual = actualCards.size();
+//
+//        Assert.assertEquals(expected, actual);
+//    }
+//
+//
+//    @Test
+//    void testToString() {
+//        Cards card = new Cards(Rank.ACE, Suit.CLUBS, 14);
+//        int expected = 52;
+//        ArrayList<Cards> actualCards = card.createDeck();
+//        System.out.println(actualCards);
+//    }
+//
+//    @Test
+//    void testShuffle() {
+//        Cards card = new Cards(Rank.ACE, Suit.CLUBS, 14);
+//        ArrayList<Cards> actualCards = card.createDeck();
+//        Collections.shuffle(actualCards);
+//        System.out.println(actualCards);
+//    }
 
 }
